Enable using Eigen Sparse Vector (#1)
* Add sparse matrix based on Eigen

* Use sparse Eigen Matrix

* Fix some debug errors

* Template gpgpu sne compute

* Template gpgpu sne raster

* Template gradient descent tsne texture

* Template hd joint prob generator

* Template hierarchical sne

* Template io

* Template sparse tsne

* Template graph algorithms

* Template embedding and add type trait headers consistently

* add knn utils file

* knn search in single function - less duplicate code

* Fix some template conversion warnings

* Consistent naming

* Fix template linkage errors

* Fix printf warning

* Fix out of bounds bug

* Fix some warnings

* Fix some warnings

* Fix some warnings

* Fix some warnings

* Fix some warnings

* remove FLANN parameters and rename knn parameters

* MSVC: set cpp version correctly and enforce standard

* Update readme: add eigen dependency

* Set search ef parameter for hnsw

* Fix typo in readme

+++ hdi/data/histogram_inl.h
@@ -89,7 +89,7 @@ namespace hdi{
         ++_histogram[_num_buckets-1];
         return;
       }
-      int id = (v-_min)/(_max-_min)*num_buckets();
+      int id = static_cast<int>((v-_min)/(_max-_min)*num_buckets());
       ++_histogram[id];
     }
